Bring back RetryPolicy for paginator
All the http requests made with GenomicsFactory have backoff handling
for IOExceptions and backend errors. Unfortunately, IOExceptions can
still occur outside of the http request - most notably in the json
parsing of the response.

This means that we can’t rely on the http handlers to catch all the
possible retries, and so we sometimes (especially for batch jobs) need
to add in additional retry logic.

All Paginator searches now default to 3 retries.

Also note that RetryPolicy is no longer typed.