Added a behavior choice for variant shard boundaries.
Prior to this change, cluster compute jobs attempting to shard the variants data will see any records that span a shard boundary in both shards.  In some cases this might be okay, in others it is not.

Change the signature of the variants paginator so that the user makes an explicit choice as to what is appropriate for the job.