Don't set the RG tag using read.getReadGroupId()
In practice these values have different semantics. In the Readstore API, the `readGroupId` is a server-generated ID which corresponds to a read group that the user imported. Every Read in the Readstore belongs to a read group, meaning that even if the original caller imported reads without a read group, we would allocate a new one for them.

This "RG" tag will already be set everywhere we would want to consume it downstream.